Android App
A patient seeking orthodontic/ dental cosmetic care to perform a personal assessment of their smile aesthetic needs aided by a “personal virtual doctor” using the medium of 3D visualization and Virtual reality 
Automatically proposing a series of potential care solutions to the patient that they can visualize dynamically and interactively i.e. “try a series of new looks” and see themselves in the altered state in a virtual and 3D environment.
Detect and measure the affective (emotional i.e. micro and macro facial expressions) response to the various proposed treatment outcomes to better appreciate and understand the patient’s “likes and dislikes”